WIP: fixed argument checker for Norm
