+++ inc/ft_push_swap.h
@@ -9,9 +9,14 @@ typedef struct s_vars
 	char	*bin_num; //memory leeks consider on exxit
 }			t_vars;
 
+void	ft_sorting_helper(t_list **s_a, t_list **s_b, int i);
+void	ft_reverse(char *x, int begin, int end);
+int		ft_check_base(char *base, int base_len, int i, int index);
+int		ft_print_base(int nbr, char *base, int base_len);
+void	ft_print_base2(int nbr, char *base, char *str, int i);
+char	*ft_put_nbrbase(int nbr, char *base);
 int		ft_is_sorted(t_list **stack_a);
 void	ft_reverse_rotate_helper(t_list **stack_a, t_list **stack_b);
-char	*ft_put_nbrbase(int nbr, char *base);
 int		ft_print_base(int nbr, char *base, int base_len); //delete this
 int		ft_check_base(char *base, int base_len, int i, int index);
 void	ft_handle_two(t_list **stack_a);
